Updated to zig 0.15.2 and some improvements
Added script for building the example bineries.
Updated the .gitignore to include the build binaries.

Added signals example binary.
signals:
Will create a child process and then will notify his parent that it was initializated.
When the parent receives the initialization message it will send a SIGUSR1 to the child.
The parent will wait for a message from the child, that the SIGUSR1 was received.
The parent will only wait for 1 second in both cases.